Rename basic stats UQP for consistency
